refactor(server): move armResultTimeoutForTest to tests/test-helpers.js
Drop the test-only helper from the production export surface of
src/sdk-session.js. PR #2859 (issue #2843) moved it out of the
SdkSession class but kept it as a module-level export — leaving it
reachable from any non-test consumer of @chroxy/server.

Moves the helper into packages/server/tests/test-helpers.js alongside
the other test utilities and updates the sole import site in
tests/sdk-session-timeout-pause.test.js. Production module now exports
only production API.

Closes #2870
